Rename / Refactor files and fields from example plugin
