+++ runelite-plugin.properties
@@ -1,6 +1,6 @@
-displayName=Example
-author=Nobody
+displayName=Arceuus Timers
+author=Mantic
 support=
-description=An example greeter plugin
-tags=
-plugins=com.example.ExamplePlugin
+description=Better timers for Arceuus Spellbook
+tags=timer,arceuus,spellbook
+plugins=com.arceuustimers.ArceuusTimersPlugin
