+++ src/test/java/com/arceuustimers/ArceuusTimersPluginTest.java
@@ -1,13 +1,13 @@
-package com.example;
+package com.arceuustimers;
 
 import net.runelite.client.RuneLite;
 import net.runelite.client.externalplugins.ExternalPluginManager;
 
-public class ExamplePluginTest
+public class ArceuusTimersPluginTest
 {
 	public static void main(String[] args) throws Exception
 	{
-		ExternalPluginManager.loadBuiltin(ExamplePlugin.class);
+		ExternalPluginManager.loadBuiltin(ArceuusTimersPlugin.class);
 		RuneLite.main(args);
 	}
 }
